Clean up FormDataHandler

+++ src/main/java/com/matthewglover/request_handler/FormDataHandler.java
@@ -20,33 +20,47 @@ public class FormDataHandler extends RequestHandler {
 
     @Override
     public HttpResponse getResponse(HttpRequest request) {
-        if (request.getMethod() == HttpRequestMethod.GET) {
-            return handleGetRequest(request);
-        } else if (request.getMethod() == HttpRequestMethod.DELETE) {
-            return handleDeleteRequest();
-        } else {
-            return handleUpdateRequest(request);
-        }
+        if (isGet(request)) return handleGetRequest();
+        if (isDelete(request)) return handleDeleteRequest();
+        else return handleUpdateRequest(request);
     }
 
-    private HttpResponse handleDeleteRequest() {
-        temporalState = null;
-        return HttpResponseFactory.get(HttpResponseTemplate.OK);
+    private boolean isGet(HttpRequest request) {
+        return request.getMethod() == HttpRequestMethod.GET;
     }
 
-    private HttpResponse handleUpdateRequest(HttpRequest request) {
-        if (request.hasContent()) {
-            temporalState = request.getContent();
-        }
-        return HttpResponseFactory.get(HttpResponseTemplate.OK);
-    }
-
-    private HttpResponse handleGetRequest(HttpRequest request) {
+    private HttpResponse handleGetRequest() {
         HttpResponse response = HttpResponseFactory.get(HttpResponseTemplate.OK);
-        if (temporalState != null) {
+        if (hasTemporalState()) {
             response.setContent(temporalState);
             response.setContentLengthHeader();
         }
         return response;
+    }
+
+    private boolean hasTemporalState() {
+        return temporalState != null;
+    }
+
+    private boolean isDelete(HttpRequest request) {
+        return request.getMethod() == HttpRequestMethod.DELETE;
+    }
+
+    private HttpResponse handleDeleteRequest() {
+        clearTemporalState();
+        return HttpResponseFactory.get(HttpResponseTemplate.OK);
+    }
+
+    private String clearTemporalState() {
+        return temporalState = null;
+    }
+
+    private HttpResponse handleUpdateRequest(HttpRequest request) {
+        if (request.hasContent()) setTemporalState(request.getContent());
+        return HttpResponseFactory.get(HttpResponseTemplate.OK);
+    }
+
+    private void setTemporalState(String newState) {
+        temporalState = newState;
     }
 }
